requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.32
sqlalchemy>=2.0
httpx>=0.27
feedparser>=6.0
apscheduler>=3.10
python-dotenv>=1.0
textblob>=0.18
scikit-learn>=1.5
jinja2>=3.1
aiofiles>=24.1
